updating task parser to order content better (more authentically VS what was done in SimYoung) moving some implementation to its own file (platform_utils.cpp)
